Sprint 4

+++ controllers/city.js
@@ -35,7 +35,7 @@ const controller = {
             };
         }
 
-        if (req.query.userId){
+        if (req.query.userId) {
             query = {
                 userId: req.query.userId
             }
@@ -43,7 +43,7 @@ const controller = {
 
         try {
             let allCities = await City.find(query);
-            if (allCities.length>0) {
+            if (allCities.length > 0) {
                 res.status(200).json({
                     success: true,
                     message: 'All cities',
@@ -69,6 +69,7 @@ const controller = {
         let { id } = req.params
 
         try {
+            
             let city = await City.findOne({ _id: id }).populate({ path: 'userId', select: 'name photo -_id' });
 
             if (city) {
@@ -95,20 +96,28 @@ const controller = {
         let { id } = req.params;
 
         try {
-            let city = await City.findOneAndUpdate({ _id: id }, req.body, { new: true });
-            if(city) {
-                res.status(200).json({
-                    success: true,
-                    message: 'City updated',
-                    data: city,
-                });
-            }else{
-                res.status(404).json({
+            let oneCity = await City.findById(id)
+            if (oneCity.userId.equals(req.user.id)) {
+                let city = await City.findOneAndUpdate({ _id: id }, req.body, { new: true });
+                if (city) {
+                    res.status(200).json({
+                        success: true,
+                        message: 'City updated',
+                        data: city,
+                    });
+                } else {
+                    res.status(404).json({
+                        success: false,
+                        message: 'City not found',
+                    });
+                }
+            } else {
+                res.status(401).json({
                     success: false,
-                    message: 'City not found',
+                    message: 'Unauthorized',
                 });
             }
-        }catch (error) {
+        } catch (error) {
             res.status(400).json({
                 success: false,
                 message: error.message,
@@ -120,20 +129,28 @@ const controller = {
         let { id } = req.params;
 
         try {
-            let city = await City.findOneAndDelete({ _id: id });
-            if(city) {
-                res.status(200).json({
-                    success: true,
-                    message: 'City deleted',
-                    data: city,
-                });
-            }else{
-                res.status(404).json({
+            let oneCity = await City.findById(id)
+            if (oneCity.userId.equals(req.user.id)) {
+                let city = await City.findOneAndDelete({ _id: id });
+                if (city) {
+                    res.status(200).json({
+                        success: true,
+                        message: 'City deleted',
+                        data: city,
+                    });
+                } else {
+                    res.status(404).json({
+                        success: false,
+                        message: 'City not found',
+                    });
+                }
+            } else {
+                res.status(401).json({
                     success: false,
-                    message: 'City not found',
+                    message: 'Unauthorized',
                 });
             }
-    }catch (error) {
+        } catch (error) {
             res.status(400).json({
                 success: false,
                 message: error.message,
