+++ controllers/hotel.js
@@ -1,7 +1,7 @@
-const { query } = require('express');
 const Hotel = require('../models/Hotel');
 
 const controller = {
+
     create: async (req, res) => {
         try {
             let new_hotel = await Hotel.create(req.body);
@@ -32,7 +32,7 @@ const controller = {
         if (req.query.order) {
             order = { name: req.query.order }
         }
-        if (req.query.userId){
+        if (req.query.userId) {
             query = {
                 userId: req.query.userId
             }
@@ -92,17 +92,25 @@ const controller = {
         let { id } = req.params;
 
         try {
-            let oneHotel = await Hotel.findOneAndUpdate({ _id: id }, req.body, { new: true });
-            if (oneHotel) {
-                res.status(200).json({
-                    success: true,
-                    message: 'Hotel updated succesfully',
-                    data: oneHotel,
-                });
+            let oneHotelFind = await Hotel.findById(id)
+            if (oneHotelFind.userId.equals(req.user.id)) {
+                let oneHotel = await Hotel.findOneAndUpdate({ _id: id }, req.body, { new: true });
+                if (oneHotel) {
+                    res.status(200).json({
+                        success: true,
+                        message: 'Hotel updated succesfully',
+                        data: oneHotel,
+                    });
+                } else {
+                    res.status(404).json({
+                        success: false,
+                        message: 'Hotel not found',
+                    });
+                }
             } else {
-                res.status(404).json({
+                res.status(403).json({
                     success: false,
-                    message: 'Hotel not found',
+                    message: "You can't update this hotel",
                 });
             }
         } catch (error) {
@@ -117,20 +125,28 @@ const controller = {
         let { id } = req.params;
 
         try {
-            let hotel = await Hotel.findOneAndDelete({ _id: id });
-            if(hotel){
-                res.status(200).json({
-                    success: true,
-                    message: 'Hotel deleted',
-                    data: hotel,
-                });
-            }else{
-                res.status(404).json({
+            let oneHotelFind = await Hotel.findById(id)
+            if (oneHotelFind.userId.equals(req.user.id)) {
+                let hotel = await Hotel.findOneAndDelete({ _id: id });
+                if (hotel) {
+                    res.status(200).json({
+                        success: true,
+                        message: 'Hotel deleted',
+                        data: hotel,
+                    });
+                } else {
+                    res.status(404).json({
+                        success: false,
+                        message: 'Hotel not found',
+                    });
+                }
+            } else {
+                res.status(403).json({
                     success: false,
-                    message: 'Hotel not found',
+                    message: "You can't delete this hotel",
                 });
             }
-    }catch (error) {
+        } catch (error) {
             res.status(400).json({
                 success: false,
                 message: error.message,
@@ -139,10 +155,4 @@ const controller = {
     },
 }
 
-
 module.exports = controller;
-
-
-
-
-
